requirements.txt
django==3.2.9
markdown
django-tinymce
gunicorn
mysqlclient
django-environ
elasticsearch>=7,<8
django-haystack
elasticsearch-dsl>=7,<8
requests
thefuzz
unidecode
markdownify
